move client to batch interface
