+++ ramls/source-record-storage-stream.raml
@@ -15,7 +15,6 @@ uses:
 types:
   errors: !include raml-storage/raml-util/schemas/errors.schema
   marcRecordSearchRequest: !include raml-storage/schemas/dto/sourceRecordSearchRequest.json
-  marcBibCollection: !include marcBibCollection.json
 
 traits:
   validate: !include raml-storage/raml-util/traits/validation.raml
@@ -52,31 +51,6 @@ traits:
           type: string[]
           example: ["order,ASC"]
           required: false
-  /verify:
-     displayName: Verify Marc Holdings Records 004 field
-     description: Verify Marc Holdings Records 004 field
-     post:
-       description: Get a list of invalid Marc Bib Record IDs, which doesn't exists in the system
-       is: [validate]
-       body:
-         application/json:
-           description: List of Marc Bib Record IDs
-           type: string[]
-       responses:
-         200:
-           body:
-             application/json:
-               type: marcBibCollection
-         400:
-           description: Bad request
-           body:
-             text/plain:
-               example: "Bad request"
-         500:
-           description: Internal server error
-           body:
-             text/plain:
-               example: "Internal server error"
   /source-records:
     displayName: Source Records Stream
     description: Stream collection of source records
